+++ app/Models/User.php
@@ -68,6 +68,17 @@ class User extends Authenticatable
         'email_verified_at' => 'datetime',
     ];
 
+    /**
+     * Route notifications for the Vonage channel.
+     *
+     * @param  \Illuminate\Notifications\Notification  $notification
+     * @return string
+     */
+    public function routeNotificationForVonage($notification)
+    {
+        return $this->mobile_number;
+    }
+
     public function offers(): HasMany
     {
         return $this->hasMany(Offer::class);