+++ app/Notifications/VerificationPinRequested.php
@@ -1,0 +1,45 @@
+<?php
+
+namespace App\Notifications;
+
+use Illuminate\Bus\Queueable;
+use Illuminate\Notifications\Messages\VonageMessage;
+use Illuminate\Notifications\Notification;
+
+class VerificationPinRequested extends Notification
+{
+    use Queueable;
+
+    /**
+     * Create a new notification instance.
+     *
+     * @return void
+     */
+    public function __construct()
+    {
+    }
+
+    /**
+     * Get the notification's delivery channels.
+     *
+     * @param  mixed  $notifiable
+     * @return array<string>
+     */
+    public function via($notifiable): array
+    {
+        return ['vonage'];
+    }
+
+    /**
+     * Get the Vonage / SMS representation of the notification.
+     *
+     * @param  mixed  $notifiable
+     * @return \Illuminate\Notifications\Messages\VonageMessage
+     */
+    public function toVonage($notifiable)
+    {
+        return (new VonageMessage)
+            ->clientReference((string) $notifiable->id)
+            ->content('Your Helping Homes verification code: '.$notifiable->verify_code);
+    }
+}