+++ database/migrations/2022_10_18_060422_remove_extra_contact_fields_from_user_table.php
@@ -1,0 +1,36 @@
+<?php
+
+use Illuminate\Database\Migrations\Migration;
+use Illuminate\Database\Schema\Blueprint;
+use Illuminate\Support\Facades\Schema;
+
+return new class extends Migration
+{
+    /**
+     * Run the migrations.
+     *
+     * @return void
+     */
+    public function up()
+    {
+        Schema::table('users', function (Blueprint $table) {
+            $table->dropColumn([
+                'alt_mobile_number',
+                'alt_mobile_number_verified_at',
+            ]);
+        });
+    }
+
+    /**
+     * Reverse the migrations.
+     *
+     * @return void
+     */
+    public function down()
+    {
+        Schema::table('users', function (Blueprint $table) {
+            $table->string('alt_mobile_number')->nullable();
+            $table->datetime('alt_mobile_number_verified_at')->nullable();
+        });
+    }
+};