+++ routes/web.php
@@ -3,17 +3,13 @@
 use App\Http\Controllers\AuthController;
 use App\Http\Controllers\OfferController;
 use App\Http\Controllers\PointOfInterestController;
+use App\Http\Controllers\VerifyController;
 use Illuminate\Support\Facades\Route;
 
 /*
 |--------------------------------------------------------------------------
 | Web Routes
 |--------------------------------------------------------------------------
-|
-| Here is where you can register web routes for your application. These
-| routes are loaded by the RouteServiceProvider within a group which
-| contains the "web" middleware group. Now create something great!
-|
 */
 
 Route::inertia('/', 'home')->name('home');
@@ -26,6 +22,9 @@ Route::apiResources([
     'offers' => OfferController::class,
     'points-of-interest' => PointOfInterestController::class,
 ]);
+
+Route::post('/verify', [VerifyController::class, 'store']);
+Route::post('/verify-code', [VerifyController::class, 'verify']);
 
 /*
 |--------------------------------------------------------------------------